Add files via upload
This code page is accessed by clicking a URL (Formula Rich Text field, HTML button) when creating a timecard record in QuickBase. When a user visits this code page, their personal/employee, and other details are sent through as URL parameters to the code page. On this page, the user has their GPS coordinates extracted and a photo is taken, and these details (and more) are sent back to QuickBase and they are officially clocked out for that timecard.