Add files via upload
This code page allows management to view the GPS coordinates of an employee, including the location where they clocked in, where they clocked out, and the location of the Project (GPS coordinates of the Project Address). You can toggle between turning on and off the clock-in and clock-out coordinates, as well as the visual Geo-Fence (1KM radius from the job site). There is an added feature with the top-right navigation menu, where you can click on the different tabs to copy the content to your clipboard.